+++ CHANGELOG.md
@@ -6,6 +6,7 @@
  - Added Tooltip
  - Added Table rowsClass prop
  - Refactored Table fields definition and selectable feature
+ - Refactored Dropdown with Popper.js
 
 ### 0.6.0